+++ Pinta.Effects/Effects/AddNoiseEffect.cs
@@ -9,6 +9,7 @@
 
 using System;
 using System.Collections.Immutable;
+using System.Threading.Tasks;
 using Cairo;
 using Pinta.Core;
 using Pinta.Gui.Widgets;
@@ -42,7 +43,7 @@ public sealed class AddNoiseEffect : BaseEffect
 		EffectData = new NoiseData ();
 	}
 
-	public override void LaunchConfiguration ()
+	public override Task<bool> LaunchConfiguration ()
 		=> chrome.LaunchSimpleEffectDialog (this);
 
 	#region Algorithm Code Ported From PDN