+++ Pinta.Effects/Effects/TwistEffect.cs
@@ -9,6 +9,7 @@
 
 using System;
 using System.Collections.Immutable;
+using System.Threading.Tasks;
 using Cairo;
 using Pinta.Core;
 using Pinta.Gui.Widgets;
@@ -38,7 +39,7 @@ public sealed class TwistEffect : BaseEffect
 		EffectData = new TwistData ();
 	}
 
-	public override void LaunchConfiguration ()
+	public override Task<bool> LaunchConfiguration ()
 		=> chrome.LaunchSimpleEffectDialog (this);
 
 	#region Algorithm Code Ported From PDN